refactor: use uint32 for AVS IDs instead of uint64

+++ x/avs/types/query.go
@@ -17,7 +17,7 @@ func NewQueryServicesRequest(pagination *query.PageRequest) *QueryServicesReques
 }
 
 // NewQueryServiceRequest creates a new QueryServiceRequest instance
-func NewQueryServiceRequest(serviceID uint64) *QueryServiceRequest {
+func NewQueryServiceRequest(serviceID uint32) *QueryServiceRequest {
 	return &QueryServiceRequest{
 		ServiceID: serviceID,
 	}
